Moves most Results functionality to BaseResults
